refactor: 更改实现方法更好维护
