+++ src/BootstrapBlazor/Components/Table/Table.razor.cs
@@ -1139,7 +1139,7 @@ public partial class Table<TItem> : ITable, IModelEqualityComparer<TItem> where 
         // 如果 TItem 是动态数据类型添加到自动清理任务中
         if (IsDataTableDynamicContext)
         {
-            ChangeDetectionCleanTask.Rent();
+            ChangeDetectionCleanTask.Rent(this);
         }
 
         if (!_firstRender)
@@ -2056,7 +2056,7 @@ public partial class Table<TItem> : ITable, IModelEqualityComparer<TItem> where 
         {
             if (IsDataTableDynamicContext)
             {
-                ChangeDetectionCleanTask.Release();
+                ChangeDetectionCleanTask.Release(this);
             }
 
             AutoRefreshCancelTokenSource?.Cancel();
