+++ test/UnitTest/Dynamic/ChangeDetectionCleanTaskTest.cs
@@ -3,53 +3,43 @@
 // See the LICENSE file in the project root for more information.
 // Maintainer: Argo Zhang(argo@live.ca) Website: https://www.blazor.zone
 
+using System.Collections.Concurrent;
 using System.Reflection;
 
-namespace UnitTest.Utils;
+namespace UnitTest.Dynamic;
 
 public class ChangeDetectionCleanTaskTest
 {
     [Fact]
     public void Release_Ok()
     {
+        var table = new MockTable();
         var type = typeof(Table<>).Assembly.GetType("BootstrapBlazor.Components.ChangeDetectionCleanTask");
         Assert.NotNull(type);
 
-        var fieldInfo = type.GetField("_tableCount", BindingFlags.NonPublic | BindingFlags.Static);
+        var fieldInfo = type.GetField("_tableCache", BindingFlags.NonPublic | BindingFlags.Static);
         Assert.NotNull(fieldInfo);
-        var v = Assert.IsType<long>(fieldInfo.GetValue(null));
+        var v = Assert.IsType<ConcurrentDictionary<ITable, byte>>(fieldInfo.GetValue(null));
 
         var methodInfo = type.GetMethod("Rent", BindingFlags.Public | BindingFlags.Static);
         Assert.NotNull(methodInfo);
-        methodInfo.Invoke(null, null);
-
-        var v1 = Assert.IsType<long>(fieldInfo.GetValue(null));
-
-        Assert.Equal(v + 1, v1);
+        methodInfo.Invoke(null, new object[] { table });
 
         // 多次调用 Rent，确保计数正确增加
-        methodInfo.Invoke(null, null);
-        v1 = Assert.IsType<long>(fieldInfo.GetValue(null));
-        Assert.Equal(v + 2, v1);
+        methodInfo.Invoke(null, new object[] { table });
+
+        // 同一个 Table 多次调用 Rent，只有一个键值
+        Assert.Single(v.Keys);
 
         methodInfo = type.GetMethod("Release", BindingFlags.Public | BindingFlags.Static);
         Assert.NotNull(methodInfo);
 
-        methodInfo.Invoke(null, null);
-        v1 = Assert.IsType<long>(fieldInfo.GetValue(null));
-        Assert.Equal(v + 1, v1);
+        methodInfo.Invoke(null, new object[] { table });
+        Assert.Empty(v.Keys);
 
         // 继续调用 Release，确保不会出现负数
-        methodInfo.Invoke(null, null);
-        v1 = Assert.IsType<long>(fieldInfo.GetValue(null));
-        Assert.Equal(v, v1);
-
-        fieldInfo.SetValue(null, 0L);
-
-        // 继续调用 Release，确保不会出现负数
-        methodInfo.Invoke(null, null);
-        v1 = Assert.IsType<long>(fieldInfo.GetValue(null));
-        Assert.Equal(0L, v1);
+        methodInfo.Invoke(null, new object[] { table });
+        Assert.Empty(v.Keys);
     }
 
     [Fact]
@@ -92,5 +82,16 @@ public class ChangeDetectionCleanTaskTest
         // _cancellationTokenSource 应该被取消
         cancellationTokenSource = fieldInfo.GetValue(null) as CancellationTokenSource;
         Assert.Null(cancellationTokenSource);
+    }
+
+    class MockTable : ITable
+    {
+        public Dictionary<string, IFilterAction> Filters { get; } = [];
+
+        public Func<Task>? OnFilterAsync { get; }
+
+        public List<ITableColumn> Columns { get; } = [];
+
+        public IEnumerable<ITableColumn> GetVisibleColumns() => [];
     }
 }
